Completed Updating the packages
- Added all 4 packages in (Release and Debug in x86 and x64)
- Turned off native debugging as it was causing an issue on window close in release (not really a problem but it was confusing and caused a false positive for issues)
- Adjusted the sample to use the packages (the read me will need to be updated later)
- Adjusted the sample to call the api controller sample controller  instead of test
-  Added a window close function to clean up the controller after window close not immediately after open
- Adjusted the Export and import stuff to work with x64 (this breaks x86, I will fix this tomorrow)
- Cleaned up a warning in x64 for a bad cast by casting a size_t to a DWORD in the buffer write
- Adjusted the nuget package files

+++ NUGET/UsbApiUtility/UsbApiUtility/UsbDevice.cpp
@@ -27,7 +27,7 @@ HANDLE ReadHandle(void* handle)
 }
 
 /// <summary>
-/// 
+/// This checks the COM device handle and then reads the device settings and returns them
 /// </summary>
 /// <param name="deviceHandle"></param>
 /// <returns></returns>
@@ -117,7 +117,7 @@ void UsbDevice::WriteDataToDevice(const char* data, size_t dataSize)
 	DWORD dwWriteBuffer = 0;
 	CheckDeviceCommState();
 	auto deviceHandle = ReadHandle(_deviceHandle);
-	if (!WriteFile(deviceHandle, data, dataSize, &dwWriteBuffer, NULL))
+	if (!WriteFile(deviceHandle, data, static_cast<DWORD>(dataSize), &dwWriteBuffer, NULL))
 	{
 		OutputDebugString(L"\nFailed to write to the USB Serial buffer \n");
 		throw UsbDeviceException(UsbDeviceError::FailedToWriteBuffer);
